+++ models/tournament.js
@@ -19,7 +19,8 @@ var schema = new Schema({
             timeStart: {type: String},
             timeDuration: {type: String},
             map: {type: String},
-            event: {type: Schema.Types.ObjectId, ref: "event"}
+            event: {type: Schema.Types.ObjectId, ref: "event"},
+            voteRoom : {type: Number}
         }],
         swiss : {
             secondRound : {type: Boolean},