+++ src/app/pages/textBox/textBox.module.ts
@@ -1,0 +1,22 @@
+import { NgModule }      from '@angular/core';
+import { CommonModule }  from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { NgaModule } from '../../theme/nga.module';
+
+import { TextBoxComponent } from './textBox.component';
+import { routing }       from './textBox.routing';
+
+
+@NgModule({
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    FormsModule,
+    NgaModule,
+    routing
+  ],
+  declarations: [
+    TextBoxComponent
+  ]
+})
+export default class TextBoxModule {}